Added CLI commands for each action listed in the README.md
clicmd.js holds all the functions called by cli.js including approveing/disapproveing nodes, creating a post, commenting, voting, transfering tokens, and updating the user profile. This allows the CLI user ignore the formatting for items such as the type number. The formatting is hardcoded and all the user needs to worry about are the string or floats that change with each transaction.